Updated JS
Created array with available spaces left (Ultimately used for computer to make move, I suppose I could use that to prevent clicks in divs with letters currently in it). 
When div is clicked, letter is placed, and that div ID (i.e. location of div) is removed from array, removing it from available spaces.